requirements.txt
streamlit
streamlit-folium
numpy
pandas
scikit-learn
tensorflow-cpu
folium
